LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGAnimatedEnumeration-SVGMarkerElement-expected.txt


PASS Test SVGAnimatedEnumeration
PASS Test SVGOrient
PASS Test grad units
FAIL Test turn units assert_equals: expected 360 but got 0
FAIL Test auto-start-reverse Type error
